Handle REST actions, and be more efficient
If the action about to run is a Catalyst::Action::REST, then we want to
look for a type-specific action that it's going to forward to, and look
at the attributes for that instead.  For e.g. if `$c->action` is the
action for `index()` in a controller and it's a `Catalyst::Action::REST`
action, then see if the controller has an e.g. `index_POST()` action,
and if so, look at the attributes for that instead.

If Catalyst::Controller::REST is in use, you may well have a thing()
action that does default stuff for interacting with a `thing`, then a
`thing_GET()` action which will obviously not expect file uploads, and a
`thing_POST()` which does expect file uploads - so if we're going to end
up at the latter, that's where we want to look for our attribute.

Also, handle multiple file type specs more efficiently, don't loop all
types listed in the attribute for each upload, assemble a hash for quick
lookup - and understand splitting on spaces/commas.